+++ src/main/kotlin/io/github/jamalam/Bot.kt
@@ -81,6 +81,12 @@ suspend fun main() {
                 }
             }
         }
+
+        hooks {
+            beforeStart {
+                logger.info { "Connecting to Discord..." }
+            }
+        }
     }
 
     bot.start()